requirements.txt
tensorflow-gpu==1.8.0
tqdm==4.23.3
matplotlib
